requirements.txt
customtkinter==5.2.2
requests==2.31.0
Pillow==10.3.0